Older versions of numpy weren't python version neutral
