+++ .travis.yml
@@ -6,6 +6,4 @@ python:
   - "3.4"
   - "3.5"
 install: python ./setup.py install
-script:
-  - nosetests -sv
-  - python -c "import numpy; print(numpy.__version__)"
+script: nosetests -sv
